Remove decoder destroy method
Due to no guarentee of HttpObject order, we can't call this safely
based on the end of response, so we leave that up to the decoder.

There is a risk of leaving temporary files in place if a response is
never completed, but adding hooks for cleanup is too complicated at
this stage.